+++ app/src/main/java/dev/yc/xmrworker/data/repository/SupportXmrRepository.kt
@@ -1,6 +1,6 @@
 package dev.yc.xmrworker.data.repository
 
-import dev.yc.xmrworker.ui.mineritem.MinerUiState
+import dev.yc.xmrworker.ui.myminer.mineritem.MinerUiState
 import kotlinx.coroutines.flow.Flow
 
 interface SupportXmrRepository {
