+++ app/src/main/java/dev/yc/xmrworker/ui/myminer/MinersViewModel.kt
@@ -1,0 +1,10 @@
+package dev.yc.xmrworker.ui.myminer
+
+import androidx.lifecycle.ViewModel
+import dev.yc.xmrworker.data.repository.SupportXmrRepository
+
+class MinersViewModel(
+    private val reposittory: SupportXmrRepository
+) : ViewModel() {
+
+}
