+++ app/src/main/java/dev/yc/xmrworker/ui/myminer/mineritem/MinerUiState.kt
@@ -1,4 +1,4 @@
-package dev.yc.xmrworker.ui.mineritem
+package dev.yc.xmrworker.ui.myminer.mineritem
 
 sealed interface MinerUiState {
     data class Success(val minerState: MinerState) : MinerUiState
