+++ app/src/main/res/layout/main_activity.xml
@@ -9,7 +9,7 @@
 
     <androidx.fragment.app.FragmentContainerView
         android:id="@+id/nav_host_fragment_container"
-        android:name="dev.yc.xmrworker.ui.main.MainFragment"
+        android:name="dev.yc.xmrworker.ui.myminer.MainFragment"
         android:layout_width="match_parent"
         android:layout_height="match_parent"
         app:defaultNavHost="true"
