+++ app/src/test/java/dev/yc/xmrworker/data/repository/FakeSupportXmrRepository.kt
@@ -1,0 +1,2 @@
+package dev.yc.xmrworker.ui.myminer 
+
